accessible buffer polish (#173852)
