+++ src/vs/workbench/contrib/terminal/browser/terminal.ts
@@ -1054,6 +1054,11 @@ export interface IXtermTerminal {
 	 * Returns a reverse iterator of buffer lines as strings
 	 */
 	getBufferReverseIterator(): IterableIterator<string>;
+
+	/**
+	 * Focuses the accessible buffer, updating its contents
+	 */
+	focusAccessibilityBuffer(): void;
 }
 
 export interface IInternalXtermTerminal {
